projetdepartage/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0